+++ lib/pdfrx.dart
@@ -4,3 +4,4 @@ export 'src/pdf_file_cache.dart';
 export 'src/widgets/pdf_viewer.dart';
 export 'src/widgets/pdf_viewer_params.dart';
 export 'src/widgets/pdf_viewer_scroll_thumb.dart';
+export 'src/widgets/pdf_widgets.dart';